+++ packages/serverless-compose/src/index.ts
@@ -1,6 +1,12 @@
 export {compose} from './compose'
 export {
   Context,
+  createMappingMiddleware,
+  createRecoveryMiddleware,
+  createRequestMappingMiddleware,
+  createResponseMappingMiddleware,
+  createTimingLogMiddleware,
+  createValidationMiddleware,
   Handler,
   Middleware,
   mappingMiddleware,